testing for cloudmanager.setup_db (tox, pytest)

+++ migrations/0002.create-transfer-meter.py
@@ -2,7 +2,7 @@ from yoyo import step
 
 step("""
 CREATE TABLE transfer_meter (
-  uploaded integer,
-  downloaded integer);
+  uploaded BIGINT,
+  downloaded BIGINT);
     """,
     """DROP TABLE transfer_meter""")
